+++ web/sdk/ui-worker.js
@@ -64,6 +64,7 @@ export const addSubsection = () => domUnavailable('addSubsection');
 export const labeledSubsection = () => domUnavailable('labeledSubsection');
 export const addFilePath = () => domUnavailable('addFilePath');
 export const createFileActions = () => domUnavailable('createFileActions');
+export const pinFile = () => domUnavailable('pinFile');
 export const addDiffViewer = () => domUnavailable('addDiffViewer');
 export const createDiffViewer = () => domUnavailable('createDiffViewer');
 export const openImageLightbox = () => domUnavailable('openImageLightbox');